+++ app/models/abilities/common.rb
@@ -99,7 +99,10 @@ module Abilities
         can :update, Budget::Investment,               budget: { phase: "accepting" }, author_id: user.id
         can :suggest, Budget::Investment,              budget: { phase: "accepting" }
         can :destroy, Budget::Investment,              budget: { phase: ["accepting", "reviewing"] }, author_id: user.id
-        can :vote, Budget::Investment,                 budget: { phase: "selecting" }
+        can :create, ActsAsVotable::Vote,
+          voter_id: user.id,
+          votable_type: "Budget::Investment",
+          votable: { budget: { phase: "selecting" }}
 
         can [:show, :create], Budget::Ballot,          budget: { phase: "balloting" }
         can [:create, :destroy], Budget::Ballot::Line, budget: { phase: "balloting" }
