+++ app/models/budget/investment.rb
@@ -277,7 +277,7 @@ class Budget
     def permission_problem(user)
       return :not_logged_in unless user
       return :organization  if user.organization?
-      return :not_verified  unless user.can?(:vote, Budget::Investment)
+      return :not_verified  unless user.can?(:create, ActsAsVotable::Vote)
 
       nil
     end
