+++ app/views/budgets/investments/votes/create.js.erb
