+++ app/views/management/budgets/investments/votes/create.js.erb
