+++ config/routes/budget.rb
@@ -2,12 +2,13 @@ resources :budgets, only: [:show, :index] do
   resources :groups, controller: "budgets/groups", only: [:show]
   resources :investments, controller: "budgets/investments" do
     member do
-      post :vote
       put :flag
       put :unflag
     end
 
     collection { get :suggest }
+
+    resources :votes, controller: "budgets/investments/votes", only: :create
   end
 
   resource :ballot, only: :show, controller: "budgets/ballots" do
