+++ spec/models/abilities/common_spec.rb
@@ -248,9 +248,9 @@ describe Abilities::Common do
       it { should_not be_able_to(:create, investment_in_selecting_budget) }
       it { should_not be_able_to(:create, investment_in_balloting_budget) }
 
-      it { should be_able_to(:vote, investment_in_selecting_budget) }
-      it { should_not be_able_to(:vote, investment_in_accepting_budget) }
-      it { should_not be_able_to(:vote, investment_in_balloting_budget) }
+      it { should be_able_to(:create, user.votes.build(votable: investment_in_selecting_budget)) }
+      it { should_not be_able_to(:create, user.votes.build(votable: investment_in_accepting_budget)) }
+      it { should_not be_able_to(:create, user.votes.build(votable: investment_in_balloting_budget)) }
 
       it { should_not be_able_to(:destroy, investment_in_accepting_budget) }
       it { should_not be_able_to(:destroy, investment_in_reviewing_budget) }
